Add a hardware type enum
The device type (RM1, RM2, RMPP) was determined in multiple places
across the code based on either the contents of the
/sys/devices/soc0/machine file, or from the user-provided --hardware
argument. This made it error-prone, and new commands or features also
needed to come with their own logic for device type determination.

This commit introduces the HardwareType enum, and its associated `parse`
function, so that the logic to determine the device type lives in a
single location.